Fixed the coreData stack, now saving works. You can now add drinks to the data and reverse them aswell. Works great! Added a singleton CoreDataManager which manages the context of the persistentContainer.

+++ Drinky/AppVariables.swift
@@ -14,6 +14,6 @@ class AppVariables {
     // the app's appdelegate
     static let appDelegate = UIApplication.shared.delegate as? AppDelegate
     
-    static var goal: Float = 2500
+    static var goal: Double = 2500
     
 }
